Circular deps fix
